only delete the test database

+++ test/ttl_tests.py
@@ -8,6 +8,7 @@ import pydocumentdb.errors as errors
 
 masterKey = '[YOUR_KEY_HERE]'
 host = '[YOUR_ENDPOINT_HERE]'
+TEST_DB_NAME = 'sample database'
 
 if masterKey == '[YOUR_KEY_HERE]' or host == '[YOUR_ENDPOINT_HERE]':
     raise Exception(
@@ -44,12 +45,13 @@ class TTLTests(unittest.TestCase):
 
         databases = list(client.ReadDatabases())
         for database in databases:
-            client.DeleteDatabase(database['_self'])
+            if database['id'] == TEST_DB_NAME:
+                client.DeleteDatabase(database['_self'])
 
     def test_collection_and_document_ttl_values(self):
         client = document_client.DocumentClient(host, {'masterKey': masterKey})
 
-        created_db = client.CreateDatabase({ 'id': 'sample database' })
+        created_db = client.CreateDatabase({ 'id': TEST_DB_NAME })
         
         collection_definition = {'id' : 'sample collection1',
                                  'defaultTtl' : 5
@@ -123,7 +125,7 @@ class TTLTests(unittest.TestCase):
     def test_document_ttl_with_positive_defaultTtl(self):
         client = document_client.DocumentClient(host, {'masterKey': masterKey})
 
-        created_db = client.CreateDatabase({ 'id': 'sample database' })
+        created_db = client.CreateDatabase({ 'id': TEST_DB_NAME })
         
         collection_definition = {'id' : 'sample collection',
                                  'defaultTtl' : 5
@@ -205,7 +207,7 @@ class TTLTests(unittest.TestCase):
     def test_document_ttl_with_negative_one_defaultTtl(self):
         client = document_client.DocumentClient(host, {'masterKey': masterKey})
 
-        created_db = client.CreateDatabase({ 'id': 'sample database' })
+        created_db = client.CreateDatabase({ 'id': TEST_DB_NAME })
         
         collection_definition = {'id' : 'sample collection',
                                  'defaultTtl' : -1
@@ -252,7 +254,7 @@ class TTLTests(unittest.TestCase):
     def test_document_ttl_with_no_defaultTtl(self):
         client = document_client.DocumentClient(host, {'masterKey': masterKey})
 
-        created_db = client.CreateDatabase({ 'id': 'sample database' })
+        created_db = client.CreateDatabase({ 'id': TEST_DB_NAME })
         
         collection_definition = {'id' : 'sample collection' }
         
@@ -279,7 +281,7 @@ class TTLTests(unittest.TestCase):
     def test_document_ttl_misc(self):
         client = document_client.DocumentClient(host, {'masterKey': masterKey})
 
-        created_db = client.CreateDatabase({ 'id': 'sample database' })
+        created_db = client.CreateDatabase({ 'id': TEST_DB_NAME })
         
         collection_definition = {'id' : 'sample collection',
                                  'defaultTtl' : 8
